shuffling functions to resolve dependency loop
